backend/requirements.txt
fastapi
uvicorn[standard]
supabase
numpy
pydantic
python-dotenv
gunicorn
pytest
slowapi
anthropic
celery[redis]
redis
pyjwt
openai>=1.0
httpx
python-multipart
psycopg2-binary
jinja2
cryptography
docusign-esign>=5.1.0
tenacity
pandas
pyarrow
scikit-learn
lightgbm
prometheus-fastapi-instrumentator==6.1.0
sendgrid==6.10.0
